backend/requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
motor==3.3.2
pymongo==4.6.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# Background tasks
celery==5.3.4
redis==5.0.1

# Google Cloud Storage
google-cloud-storage==2.13.0

# Utilities
python-dateutil==2.8.2

# PDF generation
reportlab==4.0.7

# CORS
fastapi-cors==0.0.6

